+++ server/scripts/shared.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+service_name="sysmon-server"
+sysv_source="/opt/sysmon/server/doc/sysmon-server.sh"
+config_source="/opt/sysmon/server/doc/sysmon-server.toml"
+systemd_source="/opt/sysmon/server/doc/sysmon-server.service"
+